tests/wpt/mozilla/tests/mozilla/stylesheet_load.html

<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="resources/style.css" rel="stylesheet" id="style_test"></link>
<script>
var saw_link_onload = false;
var t = async_test();
document.getElementById('style_test').onload = t.step_func(function() {
  saw_link_onload = true;
});
window.addEventListener('load', function() {
  t.step_func(function() {
    assert_true(saw_link_onload);
  });
  t.done();
}, false);
</script>
</head>
</html>
